úprava tabulky pro role, přidání dao

+++ src/main/kotlin/cz/upce/database/table/Roles.kt
@@ -3,7 +3,7 @@ package cz.upce.database.table
 import org.jetbrains.exposed.sql.Table
 
 object Roles : Table() {
-    val id = integer("id").autoIncrement()
+    val id = long("id").autoIncrement()
     val roleName = varchar("roleName", 5)
 
     override val primaryKey = PrimaryKey(id)
